Remove stale issue workflow
I saw this issue:

* https://github.com/iorate/ublacklist/issues/143

People who subscribe to this issue get a lot of spam from this GHA.
They also get more spam from people posting no-op comments just to
keep the issue open. I would subscribe to the issue because I am
interested in the feature, but I don't want the spam, so I'm not
going to. That's not a great user story.