docs(website): add Rightwing Media, Conspiracy Media, Fake-News Media blacklists (#524)
* Create Media Section; Add Rightwing Media blacklist

* Add Fake-News Media and Conspiracy Media blacklists

* Add disclaimer to Subscriptions page

Added a general disclaimer at the top of the Subscriptions page to clarify that all blacklists listed are user-generated content. The disclaimer emphasizes that the lists are not endorsed, controlled, or verified by the project, and users should exercise their own discretion when applying these blacklists. It also provides guidance for reporting any issues directly to